New utils module; restructuring of some files
